refactor: Remove scripts de extração em loop e adiciona documentação da API Pix
- Exclui arquivos antigos com chamadas em loop para 5 anos de dados
- Adiciona documentação estruturada da API transacoesPixDS em Markdown
- Organiza parâmetros e metadados para uso em notebooks Jupyter
